organize android views into packages
