+++ android/Omnivore/app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         </activity>
 
         <activity
-            android:name=".NewFlowActivity"
+            android:name=".ui.save.NewFlowActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Translucent">
             <intent-filter>
@@ -38,26 +38,5 @@
                 <data android:mimeType="text/*" />
             </intent-filter>
         </activity>
-<!--
-        <activity
-            android:name=".SaveActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.SEND" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="image/*" />
-            </intent-filter>
-            <intent-filter>
-                <action android:name="android.intent.action.SEND" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="application/pdf" />
-            </intent-filter>
-            <intent-filter>
-                <action android:name="android.intent.action.SEND" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="text/*" />
-            </intent-filter>
-        </activity>
--->
     </application>
 </manifest>
