+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/auth/GoogleSignInButton.kt
@@ -1,4 +1,4 @@
-package app.omnivore.omnivore
+package app.omnivore.omnivore.ui.auth
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.clickable
