+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/main/ScreenMain.kt
@@ -1,11 +1,13 @@
-package app.omnivore.omnivore
+package app.omnivore.omnivore.ui.main
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import app.omnivore.omnivore.screen.EmailLoginPage
-import app.omnivore.omnivore.screen.SplashPage
+import app.omnivore.omnivore.Routes
+import app.omnivore.omnivore.ui.auth.EmailLoginPage
+import app.omnivore.omnivore.ui.auth.SplashPage
+import app.omnivore.omnivore.ui.auth.LoginViewModel
 
 @Composable
 fun ScreenMain(viewModel: LoginViewModel){
