+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/save/NewFlowActivity.kt
@@ -1,6 +1,7 @@
-package app.omnivore.omnivore
+package app.omnivore.omnivore.ui.save
 
 import androidx.compose.material.ExperimentalMaterialApi
+import app.omnivore.omnivore.ui.save.SaveSheetActivity
 
 // Not sure why we need this class, but directly opening SaveSheetActivity
 // causes the app to crash.
